\name{biotypes}
\encoding{UTF-8}
\Rdversion{1.1}
\docType{methods}
\alias{biotypes,TranscriptExpressionSet-method}
\alias{biotypes}
\title{
   Accessor function for the 'biotypes' slot of a TranscriptExpressionSet object.
}

\description{
   The biotypes slot holds information on the different unique biotypes across
   all the transcripts annotated for a given gene. 
   Such information has been previously loaded by the user when creating the
   TranscriptExpressionSet object.
}

\usage{
\S4method{biotypes}{TranscriptExpressionSet}(object)
}

\arguments{
  \item{object}{a \code{TranscriptExpressionSet} object.}
}

\author{
   Mar Gonzàlez-Porta, mar@ebi.ac.uk
}
\examples{
}   
   
