% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsum.R
\name{cran_package_names}
\alias{cran_package_names}
\title{names of the packages currently available on cran}
\usage{
cran_package_names()
}
\value{
cran package names in a character vector
}
\description{
names of the packages currently available on cran
}
\details{
the function is \code{\link[memoise]{memoise}}d
}

