% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsum.R
\name{default_nsentences}
\alias{default_nsentences}
\title{default specification for the number of sentences in a paragraph}
\usage{
default_nsentences()
}
\value{
a number
}
\description{
by default a paragraph contains between 4 and 7 sentences.
}

