% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsum.R
\name{default_nwords}
\alias{default_nwords}
\title{default specification of the number of words in a sentence}
\usage{
default_nwords()
}
\value{
a number
}
\description{
by default a sentence contains between 4 and 15 words.
}

