\name{dominance}
\encoding{UTF-8}
\Rdversion{1.1}
\docType{methods}
\alias{dominance,TranscriptExpressionSet-method}
\alias{dominance}
\title{
   Accessor function for the 'dominance' slot of a TranscriptExpressionSet object.
}

\description{
   The dominance slot contains information on major transcript dominance, calculated as the ratio of expression between the second vs. first most abundant transcripts for each gene.
}

\usage{
\S4method{dominance}{TranscriptExpressionSet}(object)
}

\arguments{
  \item{object}{a \code{TranscriptExpressionSet} object.}
}

\author{
   Mar Gonzàlez-Porta, mar@ebi.ac.uk
}
\examples{
}   
   
