% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{get_cat_breeds}
\alias{get_cat_breeds}
\title{Get a list of cat breeds}
\source{
\url{http://purrfectcatbreeds.com/cat-breeds-list-with-pictures/}
}
\usage{
get_cat_breeds()
}
\description{
Get a list of cat breeds
}

