% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{get_vegetables}
\alias{get_vegetables}
\title{get a list of vegetables}
\source{
\url{http://www.vocabulary.cl/english/vegetables.htm}
}
\usage{
get_vegetables()
}
\description{
get a list of vegetables
}

