\name{gexp}
\encoding{UTF-8}
\Rdversion{1.1}
\docType{methods}
\alias{gexp,TranscriptExpressionSet-method}
\alias{gexp}
\title{
   Accessor function for the 'gexp' slot of a TranscriptExpressionSet object.
}

\description{
   The gexp slot contains a list of gene expression levels in each sample. Gene expression is calculated as the sum of transcript expression.
}

\usage{
\S4method{gexp}{TranscriptExpressionSet}(object)
}

\arguments{
  \item{object}{a \code{TranscriptExpressionSet} object.}
}

\author{
   Mar Gonzàlez-Porta, mar@ebi.ac.uk
}
\examples{
}   
   
