\name{Ipsum-package}
\alias{Ipsum-package}
\alias{Ipsum}
\docType{package}
\title{
	Ipsum: visualisation of changes in splicing in RNA-seq experiments
}
\description{
	1-5 lines of description + link to github
}
\details{
	\tabular{ll}{
		Package: \tab Ipsum\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2014-01-17\cr
		License: \tab update\cr
	}
}
\author{
	Author: Mar Gonzalez-Porta <mar@ebi.ac.uk>
}
\references{}
\keyword{ plot RNA-seq splicing }
\examples{}
