% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsum.R
\name{paragraph}
\alias{paragraph}
\title{generate a paragraph of dummy text}
\usage{
paragraph(words, latin = TRUE, lorem = TRUE,
  nsentences = default_nsentences, nwords = default_nwords)
}
\arguments{
\item{words}{custom words to use}

\item{latin}{see \code{\link{sentence}}}

\item{lorem}{should the first sentence start with "Lorem ipsum dolor amet"}

\item{nsentences}{function returning the number of sentences in a paragraph. At least 4.}

\item{nwords}{see \code{\link{sentence}}}
}
\value{
a paragraph
}
\description{
generate a paragraph (between 4 and 7 sentences).
}
\examples{
\dontrun{
paragraph( meat )
paragraph( cran_package_names() )
}
}

