\name{readBiotypeData}
\Rdversion{1.1}
\encoding{UTF-8}
\docType{methods}
\alias{readBiotypeData}

\title{
   Load transcript biotype data for a given gene from a system file.
}
\description{
   Load transcript biotype data for a given gene from a system file.
}
\usage{
readBiotypeData(gId=gId, infile=infile)
}
\arguments{
  \item{gId}{Id of the gene for which to load biotype data.}
  \item{infile}{Path to the file that contains the biotype data.}
}
\author{
   Mar Gonzàlez-Porta, <mar@ebi.ac.uk>
}
\examples{
}
