\name{rpkms}
\encoding{UTF-8}
\Rdversion{1.1}
\docType{methods}
\alias{rpkms,TranscriptExpressionSet-method}
\alias{rpkms}
\title{
   Accessor function for the 'rpkms' slot of a TranscriptExpressionSet object.
}

\description{
   The rpkms slot displays a matrix with the initial RPKM/FPKM values provided by the user. Each sample is represented as a column, and each transcript as a row.
}

\usage{
\S4method{rpkms}{TranscriptExpressionSet}(object)
}

\arguments{
  \item{object}{a \code{TranscriptExpressionSet} object.}
}

\author{
   Mar Gonzàlez-Porta, mar@ebi.ac.uk
}
\examples{
}   
   
