% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsum.R
\name{sentence}
\alias{sentence}
\title{generate a sentence of dummy words}
\usage{
sentence(words, latin = TRUE, lorem = FALSE, nwords = default_nwords)
}
\arguments{
\item{words}{custom words to use}

\item{latin}{if \code{TRUE} the specified words are mixed with pseudo latin words.}

\item{lorem}{if \code{TRUE} the sentence starts with "Lorem ipsum dolor amet"}

\item{nwords}{a function with no arguments that returns the number of words for this sentence. should return a number greater than 4.}
}
\value{
a sentence
}
\description{
generate a sentence with some 
commas placed randomly. The words used in the sentence 
are a mix of the given \code{words} and some pseudo 
latin words. The sentence starts with a capital letter 
and ends with a dot.
}
\examples{
sentence( meat, lorem = TRUE)
}

