% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsum.R
\name{prose}
\alias{prose}
\title{generate some dummy prose}
\usage{
prose(words, n, latin = TRUE, lorem = TRUE,
  nsentences = default_nsentences, nwords = default_nwords)
}
\arguments{
\item{words}{custom words}

\item{n}{number of paragraphs}

\item{latin}{see \code{\link{sentence}}}

\item{lorem}{should the first paragraph start with "Lorem ipsum dolor amet"}

\item{nsentences}{see \code{\link{paragraph}}}

\item{nwords}{see \code{\link{paragraph}}}
}
\value{
a character vector of paragraphs.
}
\description{
generate \code{n} paragraph of dummy prose
}

