% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\docType{data}
\name{filler}
\alias{filler}
\alias{meat}
\alias{vegs}
\title{Filler words}
\format{character vector}
\usage{
filler

meat

vegs
}
\description{
Filler words

Meat words

Vegetables words
}
\keyword{datasets}

